moved IsNetworkGlobalIDValid() to util.go
