+++ subnet/controllers/util.go
@@ -2,9 +2,14 @@ package controllers
 
 import (
 	"context"
+	"errors"
 
+	netGlo "gardener/networkGlobal/api/v1"
+	corev1 "gardener/subnet/api/v1"
 	v1 "gardener/subnet/api/v1"
 
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/sets"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -72,4 +77,22 @@ func (r *SubnetReconciler) deleteSubnetFinalizers() error {
 		}
 	}
 	return nil
+}
+
+func (r *SubnetReconciler) IsNetworkGlobalIDValid(obj metav1.Object) (bool, error) {
+	ctx := context.Background()
+	subnet, ok := obj.(*corev1.Subnet)
+
+	if !ok {
+		return false, errors.New("not a subnet object")
+	}
+
+	netGloID := subnet.Spec.NetworkGlobalID
+	netGlobalObject := &netGlo.NetworkGlobal{}
+
+	if err := r.Get(ctx, types.NamespacedName{Name: netGloID, Namespace: subnet.Namespace}, netGlobalObject); err != nil {
+		return false, errors.New("not valid because resource with networkGlobalID doesn't exist")
+	}
+	r.Log.Info("Succesfully got the NetWorkGlobal Object", "Name", netGlobalObject.Name)
+	return true, nil
 }
